Add YSlow bootstrapper.
YSlow can't record page load times for the first page the browser loads,
since it relies on the previous page's unload event. As a workaround,
load this page instead with a query parameter named 'url' pointing to
the real page. It will immediately redirect. Also, set
showslowIgnore=1 on the url to prevent the bootstrapper from showing
up in the results.